+++ src/olmo_core/data/__init__.py
@@ -1,14 +1,16 @@
 from .collator import DataCollator, PaddingDirection
 from .iterable_dataset import IterableDataset
 from .memmap_dataset import MemMapDataset, MemMapDatasetConfig, MemMapDType
-from .tokenizer import TokenizerConfig, TokenizerNames
+from .mixes import DataMix
+from .tokenizer import TokenizerConfig, TokenizerName
 
 __all__ = [
     "MemMapDatasetConfig",
     "MemMapDataset",
     "MemMapDType",
     "TokenizerConfig",
-    "TokenizerNames",
+    "TokenizerName",
+    "DataMix",
     "DataCollator",
     "PaddingDirection",
     "IterableDataset",