+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for unsharding model state into `safetensors` format with `olmo_core.distributed.checkpoint.unshard_checkpoint(..., use_safetensors=True)`.
-- Added `data.TokenizerConfig` config class and `data.TokenizerNames` enumeration.
+- Added `data.TokenizerConfig` config class and `data.TokenizerName` enumeration.
+- Added data mixes with `data.DataMix` API.
 
 ## [v1.0.1](https://github.com/allenai/OLMo-core/releases/tag/v1.0.1) - 2024-08-26